Move thin-client to new crate
