+++ thin-client/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "solana-thin-client"
+version = "1.12.0"
+description = "Solana Thin Client"
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+documentation = "https://docs.rs/solana-thin-client"
+edition = "2021"
+
+[dependencies]
+bincode = "1.3.3"
+log = "0.4.17"
+solana-client = { path = "../client", version = "=1.12.0" }
+solana-sdk = { path = "../sdk", version = "=1.12.0" }
+
+[dev-dependencies]
+rayon = "1.5.3"
+solana-logger = { path = "../logger", version = "=1.12.0" }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
