+++ thin-client/src/lib.rs
@@ -1,0 +1,3 @@
+#![allow(clippy::integer_arithmetic)]
+
+pub mod thin_client;
